#5 feat : 화면 이동시 전체 동의 값 리셋
