+++ client/src/components/gather/AddMilitarySavingsTerm.js
@@ -4,13 +4,39 @@ import ScrollBox from "components/common/ScrollBox";
 import SubmitButton from "components/common/SubmitButton";
 import React, { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
+import { useEffect } from "react/cjs/react.development";
 import styled from "styled-components";
 import TermCheckBox from "./TermCheckBox";
-
+const checkLists = [
+  { isCheck: false, name: "예금거래기본약관", necessary: true },
+  { isCheck: false, name: "적립식예금약관", necessary: true },
+  { isCheck: false, name: "군적금 특약", necessary: true },
+  { isCheck: false, name: "군적금 상품 설명서", necessary: true },
+  {
+    isCheck: false,
+    name: "불법/탈법 차명거래 금지 설명 확인",
+    necessary: true,
+  },
+  { isCheck: false, name: "예금자보호법 설명 확인", necessary: true },
+];
 function AddMilitarySavingsTerm() {
   const [isAllChecked, setIsAllChecked] = useState(false);
+  const [checkList, setCheckList] = useState(checkLists);
   const { state: item } = useLocation();
   const history = useNavigate();
+  useEffect(() => {
+    resetAllCheck();
+  }, []);
+  function resetAllCheck() {
+    const temp = [...checkList];
+    temp.map((item, index) => {
+      item.isCheck = false;
+
+      return item;
+    });
+    setCheckList(temp);
+    setIsAllChecked(false);
+  }
   return (
     <Container>
       <BackHeader path={-1} title={item.적금명} isScrolled={true}></BackHeader>
@@ -27,6 +53,8 @@ function AddMilitarySavingsTerm() {
           <span>를 확인합니다.</span>
         </TermPageNotice>
         <TermCheckBox
+          checkList={checkList}
+          setCheckList={setCheckList}
           isAllChecked={isAllChecked}
           setIsAllChecked={setIsAllChecked}
         ></TermCheckBox>
